requirements.txt
africastalking==1.2.7
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
contextlib2==21.6.0
Django==4.2.7
idna==3.6
python-dotenv==1.0.0
requests==2.31.0
schema==0.7.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
